+++ src/test/kotlin/com/github/hechtcarmel/jetbrainsindexmcpplugin/McpPluginUnitTest.kt
@@ -1,47 +1,14 @@
 package com.github.hechtcarmel.jetbrainsindexmcpplugin
 
-import com.github.hechtcarmel.jetbrainsindexmcpplugin.constants.JsonRpcMethods
 import com.github.hechtcarmel.jetbrainsindexmcpplugin.constants.ToolNames
 import com.github.hechtcarmel.jetbrainsindexmcpplugin.server.models.JsonRpcErrorCodes
-import com.github.hechtcarmel.jetbrainsindexmcpplugin.server.models.JsonRpcRequest
-import com.github.hechtcarmel.jetbrainsindexmcpplugin.server.models.JsonRpcResponse
 import com.github.hechtcarmel.jetbrainsindexmcpplugin.tools.ToolRegistry
 import junit.framework.TestCase
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonPrimitive
 
 class McpPluginUnitTest : TestCase() {
 
-    private val json = Json {
-        ignoreUnknownKeys = true
-        encodeDefaults = true
-    }
-
-    fun testJsonRpcRequestSerialization() {
-        val request = JsonRpcRequest(
-            id = JsonPrimitive(1),
-            method = JsonRpcMethods.TOOLS_LIST
-        )
-
-        val serialized = json.encodeToString(request)
-        val deserialized = json.decodeFromString<JsonRpcRequest>(serialized)
-
-        assertEquals(McpConstants.JSON_RPC_VERSION, deserialized.jsonrpc)
-        assertEquals(JsonRpcMethods.TOOLS_LIST, deserialized.method)
-    }
-
-    fun testJsonRpcResponseSerialization() {
-        val response = JsonRpcResponse(
-            id = JsonPrimitive(1),
-            result = JsonPrimitive("test")
-        )
-
-        val serialized = json.encodeToString(response)
-        val deserialized = json.decodeFromString<JsonRpcResponse>(serialized)
-
-        assertEquals(McpConstants.JSON_RPC_VERSION, deserialized.jsonrpc)
-        assertNull(deserialized.error)
+    fun testServerVersionTracksPluginVersion() {
+        assertEquals("5.0.0", McpConstants.SERVER_VERSION)
     }
 
     fun testToolRegistry() {